requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
PyJWT==2.9.0
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2025.2
